added some setup for inventory system
so far my intentions are setting up an inventory for every interactable that the player can access and write to and copy from to their own inventory. to display stuff,  there will be a manager object containing a script with reference lists. item data will be stored as arrays of strings