added some stuff for the dialogue managfer
need help with:
woah my goodness that is a lot of player controls...
how to fetch the comment of an Item. (btw the sound effect isnt part of the scriptable object of the item. imgonna look into that mel but eyah)
and uhhh yeah thats mostly it but wowie that controls thing is hurting